Create README.md
This repository contains automated tests for validating two scenarios related to user management in a table. The tests are implemented using Selenium and Java, with the step definitions defined in a Cucumber feature file.

Pre-requisites To run the tests, you need to have the following tools installed:

Java Development Kit (JDK) 11 Maven build tool Chrome

Running the tests To run the tests, follow these steps:

Clone this repository to your local machine (intelij) Open the Task1.feature file Put the tag on top of the scenario @wip Run your scenario acording the @tag name on MyRunner class

Test files The following files are included in this repository:

Task1.feature: contains the Cucumber feature file with the scenarios to be tested. Task1: contains the step definitions for the scenarios in task1.feature. Utilities: contains common setup and teardown methods used by the tests. pom.xml: contains the Maven configuration for the project and its dependencies.